// src/app/enum/statistics-type.ts
/**
 * 统计日期类型(当年,当月,当日，自定义年,自定义月,自定义天)
 */
export enum StatisticsType {

    /**
     * 当年
     */
    currentYear = 0,

    /**
     * 当月
     */
    currentMonth = 1,
    /**
      * 当日
      */
    currentDay = 2,
    /**
     * 自定义年
     */
    customYear = 3,

    /**
     * 自定义月
     */
    customMonth = 4,

    /**
     * 自定义日
     */
    customDay = 5,
}